+++ manufacturingorders/src/main/java/pl/hexagon/lightmom/domain/service/OrderService.java
@@ -1,7 +1,7 @@
 package pl.hexagon.lightmom.domain.service;
 
 public interface OrderService {
-    String createOrder(String orderId, Double plannedQuantity, String productId);
+    String createOrder(String orderId, Integer plannedQuantity, String productId);
 
     void completeOrder(String orderId);
 }
